mcp-client/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "mcp_client"
version = "0.1.0"
description = "MCP client for discovering and calling tools from MCP servers"
requires-python = ">=3.12"
dependencies = [
  "mcp>=1.13.1",
  "python-dotenv>=1.1.1",
  "httpx>=0.28.1",
]

[project.scripts]
mcp-client = "mcp_client.main:main"

[dependency-groups]
dev = [
  "pytest>=8.4.2",
  "ruff>=0.13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["mcp_client"]
